refactor : refactor member acceptance test(extract method)
