+++ src/test/java/nextstep/subway/member/MemberTestMethod.java
@@ -1,0 +1,109 @@
+package nextstep.subway.member;
+
+import static org.assertj.core.api.Assertions.*;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
+
+import io.restassured.RestAssured;
+import io.restassured.response.ExtractableResponse;
+import io.restassured.response.Response;
+import nextstep.subway.member.dto.MemberRequest;
+import nextstep.subway.member.dto.MemberResponse;
+
+public class MemberTestMethod {
+	public static ExtractableResponse<Response> deleteMyInformation(String token) {
+		return RestAssured
+			.given().log().all()
+			.auth().oauth2(token)
+			.contentType(MediaType.APPLICATION_JSON_VALUE)
+			.when().delete("/members/me")
+			.then().log().all()
+			.extract();
+	}
+
+	public static ExtractableResponse<Response> updateMyInformation(String token, String NEW_EMAIL, String NEW_PASSWORD, int NEW_AGE) {
+		return RestAssured
+			.given().log().all()
+			.auth().oauth2(token)
+			.contentType(MediaType.APPLICATION_JSON_VALUE)
+			.body(new MemberRequest(NEW_EMAIL, NEW_PASSWORD, NEW_AGE))
+			.when().put("/members/me")
+			.then().log().all()
+			.extract();
+	}
+
+	public static ExtractableResponse<Response> findMyInformation(String token) {
+		return RestAssured
+			.given().log().all()
+			.auth().oauth2(token)
+			.contentType(MediaType.APPLICATION_JSON_VALUE)
+			.when().get("/members/me")
+			.then().log().all()
+			.extract();
+	}
+
+	public static ExtractableResponse<Response> 회원_생성을_요청(String email, String password, Integer age) {
+		MemberRequest memberRequest = new MemberRequest(email, password, age);
+
+		return RestAssured
+			.given().log().all()
+			.contentType(MediaType.APPLICATION_JSON_VALUE)
+			.body(memberRequest)
+			.when().post("/members")
+			.then().log().all()
+			.extract();
+	}
+
+	public static ExtractableResponse<Response> 회원_정보_조회_요청(ExtractableResponse<Response> response) {
+		String uri = response.header("Location");
+
+		return RestAssured
+			.given().log().all()
+			.accept(MediaType.APPLICATION_JSON_VALUE)
+			.when().get(uri)
+			.then().log().all()
+			.extract();
+	}
+
+	public static ExtractableResponse<Response> 회원_정보_수정_요청(ExtractableResponse<Response> response, String email, String password, Integer age) {
+		String uri = response.header("Location");
+		MemberRequest memberRequest = new MemberRequest(email, password, age);
+
+		return RestAssured
+			.given().log().all()
+			.contentType(MediaType.APPLICATION_JSON_VALUE)
+			.body(memberRequest)
+			.when().put(uri)
+			.then().log().all()
+			.extract();
+	}
+
+	public static ExtractableResponse<Response> 회원_삭제_요청(ExtractableResponse<Response> response) {
+		String uri = response.header("Location");
+		return RestAssured
+			.given().log().all()
+			.when().delete(uri)
+			.then().log().all()
+			.extract();
+	}
+
+	public static void 회원_생성됨(ExtractableResponse<Response> response) {
+		assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value());
+	}
+
+	public static void 회원_정보_조회됨(ExtractableResponse<Response> response, String email, int age) {
+		MemberResponse memberResponse = response.as(MemberResponse.class);
+		assertThat(memberResponse.getId()).isNotNull();
+		assertThat(memberResponse.getEmail()).isEqualTo(email);
+		assertThat(memberResponse.getAge()).isEqualTo(age);
+	}
+
+	public static void 회원_정보_수정됨(ExtractableResponse<Response> response) {
+		assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
+	}
+
+	public static void 회원_삭제됨(ExtractableResponse<Response> response) {
+		assertThat(response.statusCode()).isEqualTo(HttpStatus.NO_CONTENT.value());
+	}
+}
